bench/requirements-gpu.txt
tabulate
tensorflow-gpu
